unimoon.sol
UNIMOON network with on chain voting, wallet liquidity locks, 10% tax:  3% auto swapped for liquidity3% added for charity, 1% auto burn.  We believe the future of the network will continue to increase as people understand the true value of decentralization.  This is the early development of an incredibly revolutionary blockchain use case that will become as common as google for a search engine.  We're very excited to help bring innovation to the space, but in the interim of developing a layer 1 solution specficially built for the decentralization of a social media  network enjoy the  advanced tokenomic structure that rewards hodlers with a growing liquidity pool.  The deflationary structure will exponentially help grow the value of each $unimoon.  Sit back and relax while head straight to the moon.